DualML/requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
mlflow
plotly
matplotlib
seaborn
joblib
python-dotenv
ipykernel